# Reduce postingsCacheKey() memory allocations