--- a/pkg/storegateway/indexcache/remote_test.go
+++ b/pkg/storegateway/indexcache/remote_test.go
@@ -762,7 +762,8 @@ func TestPostingsCacheKey_ShouldOnlyAllocateOncePerCall(t *testing.T) {
 		postingsCacheKey("user-1", blockID, lbl)
 	})
 
-	assert.InDelta(t, 1.0, actualAllocs, 0.0001)
+	// Allow for 1 extra allocation here, reported when running the test with -race.
+	assert.LessOrEqual(t, actualAllocs, 2.0)
 }
 
 func TestPostingsCacheKeyLabelHash_ShouldNotAllocateMemory(t *testing.T) {
@@ -774,7 +775,8 @@ func TestPostingsCacheKeyLabelHash_ShouldNotAllocateMemory(t *testing.T) {
 		postingsCacheKeyLabelHash(lbl)
 	})
 
-	assert.InDelta(t, 0.0, actualAllocs, 0.0001)
+	// Allow for 1 extra allocation here, reported when running the test with -race.
+	assert.LessOrEqual(t, actualAllocs, 1.0)
 }
 
 func TestPostingsCacheKeyLabelHash_ShouldBeConcurrencySafe(t *testing.T) {