@pracucci 2023-04-28
> LGTM modulo the test failure (I wasn't able to reproduce locally)

I'm not sure what's happening, but I noticed that it reports 1 extra allocation when tests run with `-race` (as in CI). No extra allocation is reported without `-race`. To move forward, I've changed the test assertion to tolerate 1 extra allocation and added a comment to explain it.